form validation on category

+++ application/views/admin/category.php
@@ -1,6 +1,15 @@
 <!-- Page Heading -->
 <div class="row">
 	<div class="col-lg-12">
+
+		<?php if(isset($error)) { ?>
+			<div class="alert alert-danger alert-link"><?=$error?></div>
+		<?php } ?>
+
+		<?php if(isset($message)) { ?>
+			<div class="alert alert-success alert-link"><?=$message?></div>
+		<?php } ?>
+
 		<div class="row">
 			<div class="col-lg-6">
 				<table id="example" class="table table-striped table-bordered" cellspacing="0" width="100%">
@@ -30,7 +39,7 @@
 									<td>
 										<div class="btn-group btn-group-sm">
 											<input form="form<?=$row->id_category?>" type="submit" value="Edit" class="btn btn-primary">
-											<a href="<?=base_url()?>admin/Category/deleteCategory/<?=$row->id_category?>" class="btn btn-danger btn-sm">
+											<a href="<?=base_url()?>admin/Category/deleteCategory/<?=$row->id_category?>" class="btn btn-danger btn-sm" onclick="return confirm('Delete Category ?')">
 												Delete
 											</a>
 										</div>
